+++ CHANGELOG.md
@@ -18,8 +18,14 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Push Notification uses `MessagingStyle` on devices with API Version 23+
+- Push Notification configuration has been simplified, check our [docs](https://getstream.io/chat/docs/sdk/android/client/guides/push-notifications/#customizing-push-notifications) to see how it works
+- `NotificationHandler` interface allows you to implement your own Push Notification logic show/remove notifications. It is the new interface you need to use if you were using `ChatNotificationHandler` previously
+- `NotificationHandlerFactory` help you to use our default `NotificationHandler` implementations
 
 ### ⚠️ Changed
+- Some properties of `NotificationConfig` has been deprecated, check our [DEPRECATIONS](https://github.com/GetStream/stream-chat-android/blob/main/DEPRECATIONS.md) section
+- `ChatNotificationhandler` class has been deprecated, you need to use `NotificationHandler` now. Check our [DEPRECATIONS](https://github.com/GetStream/stream-chat-android/blob/main/DEPRECATIONS.md) section.
 
 ### ❌ Removed
 
@@ -64,8 +70,11 @@
 
 ## stream-chat-android-ui-components
 ### 🐞 Fixed
+- Fixed bug related to scroll of messages.
 
 ### ⬆️ Improved
+- Now it is possible to set a custom `LinearLayoutManager` to `MessageListView`, this can be used to change stack of messages or revert the layout.
+- Removed full screen loading view when loading more message items on the `SearchResultListView`.
 
 ### ✅ Added
 - Added `MessageListView::getRecyclerView` method which exposes the inner `RecyclerView` with message list items.